Rename aktivitetlisteSelector og skrive om til arrow functions
